+++ src/bsp/src/stm32/hal/decawave/platform/CMakeLists.txt
@@ -12,7 +12,6 @@ set(LIB_HEADERS
         include/deca_port.h
     )
 
-
 add_library(${LIB_NAME} INTERFACE)
 add_library(${LIB_ALIAS} ALIAS ${LIB_NAME})
 
@@ -23,4 +22,15 @@ target_include_directories(${LIB_NAME}
         include
     )
 
-target_link_libraries(${LIB_NAME} INTERFACE SwarmUS::HiveMind::OS::FreeRTOS)
+if (${COMPILE_STM32_F429ZI})
+    add_subdirectory(src/stm32f429zi)
+elseif (${COMPILE_STM32_H735ZG})
+    add_subdirectory(src/stm32h735zg)
+else()
+    message(FATAL_ERROR "STM PORT: The current board is not supported" )
+endif()
+
+target_link_libraries(${LIB_NAME} INTERFACE
+        SwarmUS::HiveMind::HAL::Decawave::Platform::Device
+        SwarmUS::HiveMind::OS::FreeRTOS
+        )
