+++ src/bsp/src/stm32/hal/decawave/platform/src/deca_port.c
@@ -96,32 +96,3 @@ void deca_isr(decaDevice_t selectedDevice) {
         }
     }
 }
-
-void deca_pulseSyncSignal() {
-    // Enable sync
-    deca_setSyncEnable(true);
-    deca_setSyncClear(true);
-
-    // Sync
-    // TODO: Maybe play on timings here
-    deca_setSync(true);
-    vTaskDelay(1);
-    deca_setSync(false);
-    deca_setSyncClear(false);
-
-    // Disable sync
-    deca_setSyncEnable(false);
-}
-
-void deca_setSync(bool state) {
-    HAL_GPIO_WritePin(DW_SYNC_GPIO_Port, DW_SYNC_Pin, state ? GPIO_PIN_SET : GPIO_PIN_RESET);
-}
-
-void deca_setSyncEnable(bool state) {
-    HAL_GPIO_WritePin(DW_SYNC_EN_GPIO_Port, DW_SYNC_EN_Pin, state ? GPIO_PIN_SET : GPIO_PIN_RESET);
-}
-
-void deca_setSyncClear(bool state) {
-    HAL_GPIO_WritePin(DW_SYNC_CLEAR_GPIO_Port, DW_SYNC_CLEAR_Pin,
-                      state ? GPIO_PIN_SET : GPIO_PIN_RESET);
-}
