+++ src/bsp/src/stm32/hal/decawave/platform/src/stm32f429zi/deca_platform.c
@@ -1,4 +1,35 @@
-#include "hivemind_hal.h"
+#include "deca_port.h"
+#include <FreeRTOS.h>
+#include <task.h>
+
+void deca_setSync(bool state) {
+    HAL_GPIO_WritePin(DW_SYNC_GPIO_Port, DW_SYNC_Pin, state ? GPIO_PIN_SET : GPIO_PIN_RESET);
+}
+
+void deca_setSyncEnable(bool state) {
+    HAL_GPIO_WritePin(DW_SYNC_EN_GPIO_Port, DW_SYNC_EN_Pin, state ? GPIO_PIN_SET : GPIO_PIN_RESET);
+}
+
+void deca_setSyncClear(bool state) {
+    HAL_GPIO_WritePin(DW_SYNC_CLEAR_GPIO_Port, DW_SYNC_CLEAR_Pin,
+                      state ? GPIO_PIN_SET : GPIO_PIN_RESET);
+}
+
+void deca_pulseSyncSignal() {
+    // Enable sync
+    deca_setSyncEnable(true);
+    deca_setSyncClear(true);
+
+    // Sync
+    // TODO: Maybe play on timings here
+    deca_setSync(true);
+    vTaskDelay(1);
+    deca_setSync(false);
+    deca_setSyncClear(false);
+
+    // Disable sync
+    deca_setSyncEnable(false);
+}
 
 // TODO: Get num devices
 decawaveDeviceConfig_t g_decawaveConfigs[2] = {{.spiHandle = &hspi4,
